Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-invoke-insert-source-networkState-expected.txt


FAIL NOT invoking resource selection by inserting <source> when networkState is not NETWORK_EMPTY assert_equals: networkState in second script expected 2 but got 3
